requirements.txt
python-telegram-bot[job-queue,webhooks]==20.7
python-dotenv==1.0.0
pytz==2023.3
Flask==2.3.3
requests==2.31.0